check
start - check commit

+++ src/Light.java
@@ -2,9 +2,9 @@ public class Light {
 
     public Vector position;
     public Vector color;
-    public double specular_intencity;
-    public double shadow_intencity;
-    public double radius;
+    public float specular_intencity;
+    public float shadow_intencity;
+    public float radius;
 
 
 }
