+++ src/Material.java
@@ -3,8 +3,8 @@ public class Material {
     public Vector diffuse_color;
     public Vector spectular_color;
     public Vector reflection_color;
-    public double phong_specularity_coefficient;
-    public double transparency;
+    public float phong_specularity_coefficient;
+    public float transparency;
 
 
 }
